api\grpc\server compiles on Windows
